% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_proposal.R
\name{adaptive_proposal_control}
\alias{adaptive_proposal_control}
\title{Adaptive proposal control}
\usage{
adaptive_proposal_control(
  initial_scaling = 0.2,
  scaling_increment = 0.01,
  acceptance_target = 0.234,
  initial_weight = 1000,
  adaptive_contribution = 0.95
)
}
\arguments{
\item{initial_scaling}{The initial scaling}

\item{scaling_increment}{The scaling increment}

\item{acceptance_target}{The fraction of proposals that should be
accepted (optimally).}

\item{initial_weight}{Initial weight}

\item{adaptive_contribution}{The fractional contribution of the
adaptive part of the proposal}
}
\description{
Control for adaptive proposals, used in \link{pmcmc_control}
for deterministic models.
}
