% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{array_bind}
\alias{array_bind}
\title{Bind arrays by last dimension}
\usage{
array_bind(..., arrays = list(...))
}
\arguments{
\item{...}{Any number of arrays. All dimensions must the the same,
except for the final one (representing time) which may vary.}

\item{arrays}{As an alternative to using \code{...} you can provide a
list directly. This is often nicer to program with.}
}
\value{
A single array object
}
\description{
Bind a number of arrays by their last dimension. This is useful
for binding together the sorts of arrays produced by dust and
mcstate's simulation functions.
}
\examples{
# Consider two matricies; this is equivalent to rbind and is
# pretty trivial
m1 <- matrix(1, 4, 5)
m2 <- matrix(2, 4, 2)
mcstate::array_bind(m1, m2)

# For a 4d array though it's less obvious
a1 <- array(1, c(2, 3, 4, 5))
a2 <- array(2, c(2, 3, 4, 1))
a3 <- array(3, c(2, 3, 4, 3))
dim(mcstate::array_bind(a1, a2, a3))
}
