% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrays.R
\name{array_flatten}
\alias{array_flatten}
\title{Flatten array dimensions}
\usage{
array_flatten(x, i)
}
\arguments{
\item{x}{An array}

\item{i}{An integer vector of dimensions to flatten}
}
\value{
A new array with at one or more dimensions removed
}
\description{
Flatten array dimensions into a single dimension. This takes a
multidimensional array and converts some dimensions of it into a
vector. Use this to drop out "middle" dimensions of a structured
array. This is conceptually the inverse of \link{array_reshape}
}
\examples{
x <- array(1:12, c(2, 3, 4))
mcstate::array_flatten(x, 2:3)

# array_flatten and array_reshape are each others' conceptual
# opposites:
y <- mcstate::array_flatten(x, 2:3)
identical(mcstate::array_reshape(y, 2, c(3, 4)), x)
}
\seealso{
\link{array_flatten} which adds structure
}
