% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast.mcstate_scan}
\title{Sample Grid Scan}
\usage{
forecast(
  x,
  ...,
  filter,
  n_sample_pairs = 10,
  n_particles = 100,
  forecast_steps = 0
)

\method{forecast}{mcstate_scan}(
  x,
  ...,
  filter,
  n_sample_pairs = 10,
  n_particles = 100,
  forecast_steps = 0
)
}
\arguments{
\item{x}{Output of \code{\link{grid_search}}.}

\item{...}{Other arguments}

\item{filter}{A \code{particle_filter} object to run (the same
(as the one used to produce \code{scan_results})}

\item{n_sample_pairs}{Number of parameter pairs to be sampled. This will
determine how many trajectories are returned. Integer. Default = 10. This
will determine how many trajectories are returned.}

\item{n_particles}{Number of particles. Positive Integer. Default = 100}

\item{forecast_steps}{Number of time steps being forecast. Default = 0}
}
\value{
\code{\link{list}}. First element (trajectories) is a 3
  dimensional array of trajectories (time, state, tranjectories). Second
  element (param_grid) is the parameters chosen when sampling from the
  \code{scan_results} grid and the third dimension (inputs) is a list of
  model inputs.
}
\description{
Take a grid search produced by \code{\link{grid_search}} and
sample \code{n_sample_pairs} from the parameter grid uses based
on their probability. For each parameter pair chosen, run particle
filter with \code{num_particles}.
}
