% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\name{grid_search}
\alias{grid_search}
\title{Grid search over two parameters}
\usage{
grid_search(range, filter, tolerance = 0.01)
}
\arguments{
\item{range}{\code{data.frame} describing the parameters to search.
Must have columns name, min, max, n and target}

\item{filter}{A \code{particle_filter} object to run}

\item{tolerance}{Check around edges of the probability matrix}
}
\value{
List of beta and start date grid values, and
  normalised probabilities at each point
}
\description{
Run a grid search of the particle filter over two parameters
}
