% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if2_control.R
\name{if2_control}
\alias{if2_control}
\title{Control for IF2}
\usage{
if2_control(pars_sd, iterations, n_par_sets, cooling_target, progress = TRUE)
}
\arguments{
\item{pars_sd}{The initial standard deviation of parameter walks.}

\item{iterations}{The number of IF2 iterations to run, across which
the cooling is performed}

\item{n_par_sets}{The number of parameter sets to walk (c.f. the
population size)}

\item{cooling_target}{A factor < 1 multiplying pars_sd, which will be
reached by the end of the iterations, and approached geometrically}

\item{progress}{Logical, indicating if a progress bar should be
displayed, using \code{\link[progress:progress_bar]{progress::progress_bar}}.}
}
\value{
An \code{if2_control} object, which should not be modified once created.
}
\description{
Control for \link{if2}. This function constructs a list of options and
does some basic validation. Do not manually change the values in this
object. Do not refer to any argument by
position as the order of the arguments may change in future.
}
\examples{
mcstate::if2_control(list(beta = 0.2, gamma = 0.2), 100, 1000, 0.5)
}
