% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if2_parameters.R
\name{if2_parameters}
\alias{if2_parameters}
\title{if2_parameters}
\description{
Construct parameters for use with
\code{\link[=if2]{if2()}}. This creates a utility object that is used
internally to work with parameters. Most users only need to
construct this object, but see the examples for how it can be
used.
}
\examples{
# Construct an object with two parameters:
pars <- mcstate::if2_parameters$new(
  list(mcstate::if2_parameter("a", 0.1, min = 0, max = 1,
                                prior = function(a) log(a)),
       mcstate::if2_parameter("b", 0, prior = dnorm)))

# Initial parameters
pars$initial()

# Create the initial parameter set
n_par_sets <- 5
pars_sd <- list("a" = 0.02, "b" = 0.02)
p_mat <- pars$walk_initialise(n_par_sets, pars_sd)
p_mat

# Propose a new parameter set
p_mat <- pars$walk(p_mat, pars_sd)
p_mat

# Information about parameters:
pars$names()
pars$summary()

# Compute prior
pars$prior(p_mat)

# Transform data for your model
pars$model(p_mat)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{if2_parameters$new()}}
\item \href{#method-initial}{\code{if2_parameters$initial()}}
\item \href{#method-walk_initialise}{\code{if2_parameters$walk_initialise()}}
\item \href{#method-walk}{\code{if2_parameters$walk()}}
\item \href{#method-names}{\code{if2_parameters$names()}}
\item \href{#method-summary}{\code{if2_parameters$summary()}}
\item \href{#method-prior}{\code{if2_parameters$prior()}}
\item \href{#method-model}{\code{if2_parameters$model()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create the if2_parameters object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$new(parameters, transform = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A \code{list} of
\code{\link{if2_parameter}} objects, each of which describe a
single parameter in your model. If \code{parameters} is named, then
these names must match the \verb{$name} element of each parameter is
used (this is verified).}

\item{\code{transform}}{An optional transformation function to apply
to your parameter vector immediately before passing it to the
model function. If not given, then \link{as.list} is
used, as dust models require this. However, if you need to
generate derived parameters from those being actively sampled
you can do arbitrary transformations here.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initial"></a>}}
\if{latex}{\out{\hypertarget{method-initial}{}}}
\subsection{Method \code{initial()}}{
Return the initial parameter values as a named numeric
vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$initial()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-walk_initialise"></a>}}
\if{latex}{\out{\hypertarget{method-walk_initialise}{}}}
\subsection{Method \code{walk_initialise()}}{
Set up a parameter walk
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$walk_initialise(n_par_sets, pars_sd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_par_sets}}{An integer number of parameter sets, which
defines the size of the population being peturbed.}

\item{\code{pars_sd}}{A vector of standard deviations for the walk
of each parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-walk"></a>}}
\if{latex}{\out{\hypertarget{method-walk}{}}}
\subsection{Method \code{walk()}}{
Propose a new parameter matrix given a current matrix
and walk standard deviation vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$walk(pars, pars_sd)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{A parameter matrix, from this function or
\verb{$walk_initialise()}}

\item{\code{pars_sd}}{A vector of standard deviations for the walk
of each parameter}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-names"></a>}}
\if{latex}{\out{\hypertarget{method-names}{}}}
\subsection{Method \code{names()}}{
Return the names of the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Return a \code{\link{data.frame}} with information about
parameters (name, min, max, and integer).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prior"></a>}}
\if{latex}{\out{\hypertarget{method-prior}{}}}
\subsection{Method \code{prior()}}{
Compute the prior for a parameter vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$prior(pars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{a parameter matrix from \verb{$walk()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model"></a>}}
\if{latex}{\out{\hypertarget{method-model}{}}}
\subsection{Method \code{model()}}{
Apply the model transformation function to a parameter
vector. Output is a list for lists, suitable for use with a dust
object with \code{pars_multi = TRUE}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{if2_parameters$model(pars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{a parameter matrix from \verb{$walk()}}
}
\if{html}{\out{</div>}}
}
}
}
