% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_multistage.R
\name{multistage_epoch}
\alias{multistage_epoch}
\title{Multistage filter epoch}
\usage{
multistage_epoch(start, pars = NULL, transform_state = NULL)
}
\arguments{
\item{start}{The start time, in units of time in your data set.
These must correspond to time points with data.  The model will
complete the step to this time point, then change parameters,
then continue (so \code{start} represents the time point we move from
with these parameters)}

\item{pars}{Optional parameter object, replacing the model
parameters at this point.  If \code{NULL} then the model parameters
are not changed, and it is assumed that you will be changing
model state via \code{transform_state}.}

\item{transform_state}{Optional parameter transformation function.
This could be used in two cases (1) arbitrary change to the
model state (e.g., a one-off movement of state within particles
at a given time point that would be otherwise awkward to code
directly in your model), or (2) where you have provided \code{pars}
and these imply a different model state size.  In this case you
\emph{must} provide \code{transform_state} to fill in new model state,
move things around, or delete model state depending on how the
state has changed.  This function will be passed three
arguments: (1) the current model state, (2) the model used to
run up to this point, (3) the new model that was created with
\code{pars} which will be run from this point.  It is expected that
the results of \verb{$pars()} and \verb{$info()} from these model objects
will be useful for updating the model state.}
}
\description{
Describe an epoch within a \link{multistage_parameters} object
}
