% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_multistage.R
\name{multistage_parameters}
\alias{multistage_parameters}
\title{Multistage filter parameters}
\usage{
multistage_parameters(pars, epochs)
}
\arguments{
\item{pars}{The parameters covering the period up to the first
change in \code{epoch}.}

\item{epochs}{A list of \code{multistage_epoch} objects corresponding
to a new paramter regime starting at a new time point.}
}
\value{
An object of class \code{multistage_parameters}, suitable to
pass through to the \code{run} method of \code{\link{particle_filter}}
}
\description{
Construct parameters for a multi-stage particle filter.
}
