% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_data.R
\name{particle_filter_data}
\alias{particle_filter_data}
\title{Prepare data for use with particle filter}
\usage{
particle_filter_data(data, time, rate)
}
\arguments{
\item{data}{A \code{\link{data.frame}} of data}

\item{time}{The name of a column within \code{data} that
represents your measure of time.  This column must be
integer-like}

\item{rate}{The number of model "steps" that occur between each
time point (in \code{time}).  This must also be integer-like}
}
\value{
A data.frame with new columns \code{step_start} and
  \code{step_end} (required by \code{\link{particle_filter}}),
  along side all previous data except for the time variable, which
  is replaced by new \code{<time>_start} and \code{<time>_end}
  columns.
}
\description{
Prepare data for use with the \code{\link{particle_filter}}.  This
function is not required to use the particle filter but it helps
arrange data and be explicit about the off-by-one errors that can
occur.  It takes as input your data to compare against a model,
including some measure of "time".  We need to convert this time
into model steps.
}
\examples{
d <- data.frame(day = 5:20, y = runif(16))
mcstate::particle_filter_data(d, "day", 4)
}
\author{
Rich Fitzjohn
}
