% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter_state.R
\name{particle_filter_state}
\alias{particle_filter_state}
\title{Particle filter state}
\description{
Particle filter internal state. This object is not
ordinarily constructed directly by users, but via the
\verb{$run_begin} method to \link{particle_filter}. It provides
an advanced interface to the particle filter that allows
partially running the particle filter over part of the time
trajectory.

This state object has a number of public fields that you can read
but must not write (they are not read-only so you \emph{could} write
them, but don't).
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The dust model being simulated}

\item{\code{history}}{The particle history, if created with
\code{save_history = TRUE}. This is an internal format subject to}

\item{\code{restart_state}}{Full model state at a series of points in
time, if the model was created with non-\code{NULL} \code{save_restart}.
This is a 3d (or greater) array as described in
\link{particle_filter}}

\item{\code{log_likelihood}}{The log-likelihood so far. This starts at
0 when initialised and accumulates value for each step taken.}

\item{\code{log_likelihood_step}}{The log-likelihood attributable to the
last step (i.e., the contribution to \code{log_likelihood} made on the
last call to \verb{$step()}.}

\item{\code{current_step_index}}{The index of the last completed step.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{particle_filter_state$new()}}
\item \href{#method-run}{\code{particle_filter_state$run()}}
\item \href{#method-step}{\code{particle_filter_state$step()}}
\item \href{#method-fork_multistage}{\code{particle_filter_state$fork_multistage()}}
\item \href{#method-fork_smc2}{\code{particle_filter_state$fork_smc2()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise the particle filter state. Ordinarily
this should not be called by users, and so arguments are barely
documented.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_filter_state$new(
  pars,
  generator,
  model,
  data,
  data_split,
  steps,
  n_particles,
  has_multiple_parameters,
  n_threads,
  initial,
  index,
  compare,
  constant_log_likelihood,
  gpu_config,
  seed,
  min_log_likelihood,
  save_history,
  save_restart
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{Parameters for a single phase}

\item{\code{generator}}{A dust generator object}

\item{\code{model}}{If the generator has previously been initialised}

\item{\code{data}}{A \link{particle_filter_data} data object}

\item{\code{data_split}}{The same data as \code{data} but split by step}

\item{\code{steps}}{A matrix of step beginning and ends}

\item{\code{n_particles}}{Number of particles to use}

\item{\code{has_multiple_parameters}}{Compute multiple likelihoods at once?}

\item{\code{n_threads}}{The number of threads to use}

\item{\code{initial}}{Initial condition function (or \code{NULL})}

\item{\code{index}}{Index function (or \code{NULL})}

\item{\code{compare}}{Compare function}

\item{\code{constant_log_likelihood}}{Constant log likelihood function}

\item{\code{gpu_config}}{GPU configuration, passed to \code{generator}}

\item{\code{seed}}{Initial RNG seed}

\item{\code{min_log_likelihood}}{Early termination control}

\item{\code{save_history}}{Logical, indicating if we should save history}

\item{\code{save_restart}}{Vector of steps to save restart at}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
Run the particle filter to the end of the data. This is
a convenience function around \verb{$step()} which provides the correct
value of \code{step_index}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_filter_state$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-step"></a>}}
\if{latex}{\out{\hypertarget{method-step}{}}}
\subsection{Method \code{step()}}{
Take a step with the particle filter. This moves
the particle filter forward one step within the \emph{data} (which
may correspond to more than one step with your model) and
returns the likelihood so far.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_filter_state$step(step_index, partial = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{step_index}}{The step \emph{index} to move to. This is not the same
as the model step, nor time, so be careful (it's the index within
the data provided to the filter). It is an error to provide
a value here that is lower than the current step index, or past
the end of the data.}

\item{\code{partial}}{Logical, indicating if we should return the partial
likelihood, due to this step, rather than the full likelihood so far.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fork_multistage"></a>}}
\if{latex}{\out{\hypertarget{method-fork_multistage}{}}}
\subsection{Method \code{fork_multistage()}}{
Create a new \code{particle_filter_state} object based on
this one (same model, position in time within the data) but with
new parameters, to support the "multistage particle filter".
Unlike \code{fork_smc2}, here the parameters may imply a different
model shape and arbitrary transformations of the state are
allowed.  The model is not rerun to the current point, just
transformed at that point.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_filter_state$fork_multistage(model, pars, transform_state)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A model object (or NULL)}

\item{\code{pars}}{New model parameters}

\item{\code{transform_state}}{A function to transform the model state
from the old to the new parameter set.  See
\code{\link[=multistage_epoch]{multistage_epoch()}} for details.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fork_smc2"></a>}}
\if{latex}{\out{\hypertarget{method-fork_smc2}{}}}
\subsection{Method \code{fork_smc2()}}{
Create a new \code{particle_filter_state} object based
on this one (same model, position in time within the data) but
with new parameters, run up to the date, to support the \code{\link[=smc2]{smc2()}}
algorithm. To do this, we create a new
\code{particle_filter_state} with new parameters at the beginning of
the simulation (corresponding to the start of your data or the
\code{initial} argument to \link{particle_filter}) with your new
\code{pars}, and then run the filter foward in time until it reaches
the same step as the parent model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_filter_state$fork_smc2(pars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{New model parameters}
}
\if{html}{\out{</div>}}
}
}
}
