% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_tools.R
\name{pmcmc_combine}
\alias{pmcmc_combine}
\title{Combine pmcmc samples}
\usage{
pmcmc_combine(..., samples = list(...))
}
\arguments{
\item{...}{Arguments representing \code{\link[=pmcmc]{pmcmc()}} sample, i.e.,
\code{mcstate_pmcmc} objects. Alternatively, pass a list as the
argument \code{samples}. Names are ignored.}

\item{samples}{A list of \code{mcstate_pmcmc} objects. This is often
more convenient for programming against than \code{...}}
}
\description{
Combine multiple \code{\link[=pmcmc]{pmcmc()}} samples into one object
}
