% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{pmcmc_combine_chains}
\alias{pmcmc_combine_chains}
\title{Combine multiple chains into a single (master) chain}
\usage{
pmcmc_combine_chains(x, burn_in)
}
\arguments{
\item{x}{An mcstate_pmcmc_list containing chains to combine}

\item{burn_in}{an integer denoting the number of samples to discard
from each chain}
}
\description{
Combine multiple chains into a single (master) chain
}
