% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_parameters.R
\name{pmcmc_parameter}
\alias{pmcmc_parameter}
\title{Describe single pmcmc parameter}
\usage{
pmcmc_parameter(
  name,
  initial,
  min = -Inf,
  max = Inf,
  discrete,
  integer = FALSE,
  prior = NULL
)
}
\arguments{
\item{name}{Name for the parameter (a string)}

\item{initial}{Initial value for the parameter}

\item{min}{Optional minimum value for the parameter (otherwise
\code{-Inf}). If given, then \code{initial} must be at least this
value.}

\item{max}{Optional max value for the parameter (otherwise
\code{Inf}). If given, then \code{initial} must be at most this
value.}

\item{discrete}{Deprecated. Logical, indicating if this parameter is
discrete. If \code{TRUE} then the parameter will be rounded
after a new parameter is proposed.}

\item{integer}{Logical, indicating if this parameter is
integer. If \code{TRUE} then the parameter will be rounded
after a new parameter is proposed.}

\item{prior}{A prior function (if not given an improper flat prior
is used - be careful!). It must be a function that takes a
single argument, being the value of this parameter. If given,
then \code{prior(initial)} must evaluate to a finite value.}
}
\description{
Describe a single parameter for use within the pmcmc. Note that
the name is not set here, but will end up being naturally defined
when used with \code{\link{pmcmc_parameters}}, which collects
these together for use with \code{\link[=pmcmc]{pmcmc()}}.
}
\examples{
pmcmc_parameter("a", 0.1)
}
