% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{pmcmc_predict}
\alias{pmcmc_predict}
\title{Run predictions from PMCMC}
\usage{
pmcmc_predict(object, steps, n_threads = NULL, seed = 1L)
}
\arguments{
\item{object}{The results of running [pmcmc()] with `return_state
= TRUE` (without this extra information, prediction is not
possible)}

\item{steps}{A vector of time steps to return predictions for. The
first value must be the final value run in your simulation. An
error will be thrown if you get this value wrong, look in
`object$predict$step` (or the error message) for the correct
value.}

\item{n_threads}{The number of threads used in the simulation. If
not given, we default to the value used in the particle filter
that was used in the pmcmc.}

\item{seed}{The random number seed - a positive integer.}
}
\description{
Run predictions from the results of [pmcmc()]. This function can
also be called by running [predict()] on the object, using R's S3
dispatch.
}
