% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_tools.R
\name{pmcmc_thin}
\alias{pmcmc_thin}
\title{Thin a pmcmc chain}
\usage{
pmcmc_thin(object, burnin = NULL, thin = NULL)
}
\arguments{
\item{object}{Results of running \code{\link{pmcmc}}}

\item{burnin}{Optional integer number of iterations to discard as
"burn-in". If given then samples \code{1:burnin} will be
excluded from your results. Remember that the first sample
represents the starting point of the chain. It is an error if
this is not a positive integer or is greater than or equal to
the number of samples (i.e., there must be at least one sample
remaining after discarding burnin).}

\item{thin}{Optional integer thinning factor. If given, then every
\code{thin}'th sample is retained (e.g., if \code{thin} is 10
then we keep samples 1, 11, 21, ...).}
}
\description{
Thin results of running \code{\link{pmcmc}}. This function may be
useful before using \code{\link{pmcmc_predict}}, or before saving
pmcmc output to disk.
}
