% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_tools.R
\name{pmcmc_tidy_chains}
\alias{pmcmc_tidy_chains}
\title{Convert sampled parameters and corresponding likelihoods from
\code{\link[=pmcmc]{pmcmc()}} to a Tidy (long) format}
\usage{
pmcmc_tidy_chains(object)
}
\arguments{
\item{object}{Results of running \code{\link[=pmcmc]{pmcmc()}}.}
}
\description{
Convert sampled parameters and corresponding likeligoods from \code{\link[=pmcmc]{pmcmc()}}
to a Tidy (long) format. Helper function useful for ggplotters.
}
