% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_tune.R
\name{pmcmc_tune}
\alias{pmcmc_tune}
\title{Tune pmcmc proposals}
\usage{
pmcmc_tune(
  n_tune_blocks,
  n_tune_steps,
  pars,
  filter,
  control,
  initial = NULL,
  kernel_scaling = NULL,
  weighting_rate = 1
)
}
\arguments{
\item{n_tune_blocks}{The number of tuning blocks}

\item{n_tune_steps}{The number of steps per block}

\item{pars}{A \code{\link{pmcmc_parameters}} object containing information
about parameters (ranges, priors, proposal kernel, translation
functions for use with the particle filter). \strong{Important}: the
new proposal matrix will be set into this object (i.e.,
modifying the original)}

\item{filter}{A \code{\link{particle_filter}} object}

\item{control}{Base \link{pmcmc_control} object.  We will
override a few things in here (especially the number of steps,
trajectory information and burnin/thinning), but we'll reuse
others (parallel control, number of chains, rerun control)}

\item{kernel_scaling}{The scaling to apply to the variance
covariance matrix.  If \code{NULL} we use the "optimal" scaling
(\verb{2.38^2 / (number of parameters)})}

\item{weighting_rate}{The rate at which to reduce weight on
previous blocks of parameters for calculation of the covariance
matrix. Increasing this value towards \code{Inf} will use only the
last block, while decreasing it to 0 will use all parameters
observed so far.}
}
\value{
A list with elements
\itemize{
\item \code{initial}: new initial parameter values, suitable to pass into \code{pmcmc}
\item \code{vcv}: the final estimated variance-covariance matrix
\item \code{proposal_kernel}: the final proposal kernel (scaled appropriately)
\item \code{history}: a list of parameters and probabilities, as 4d arrays
(parameter, \code{n_tune_steps}, \code{n_chains}, \code{n_tune_blocks}).
}
}
\description{
Tune pmcmc proposals by running a series of short chains (without
saving trajectories etc) and adapting the proposal kernel to match
the scaled observed variance-covariance matrix of the parameters.
}
