% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc2.R
\name{smc2}
\alias{smc2}
\title{Run SMC^2}
\usage{
smc2(pars, filter, control)
}
\arguments{
\item{pars}{A \code{\link{smc2_parameters}} object containing
information about parameters (ranges, priors, proposal kernel,
translation functions for use with the particle filter).}

\item{filter}{A \code{\link{particle_filter}} object}

\item{control}{A \link{smc2_control} object to control the
behaviour of the algorithm}
}
\value{
A \code{smc2_result} object, with elements
\itemize{
\item \code{pars}: a matrix of sampled parameters (n_parameter_set long)
\item \code{probabilities}: a matrix of probabilities (\code{log_prior}, \code{log_likelihood},
\code{log_posterior} and \code{weight}). The latter is the log posterior
normalised over all samples
\item \code{statistics}: interesting or useful statistics about your sample,
including the \code{ess} (effective sample size, over time),
\code{acceptance_rate} (where a regeneration step was done, the acceptance
rate), \code{n_particles}, \code{n_parameter_sets} and \code{n_steps} (inputs to
the simulation). The \code{effort} field is a rough calculation of the
number of particle-filter runs that this run was worth.
}
}
\description{
Run a SMC^2. This is experimental and subject to change. Use at
your own risk.
}
