% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc2_control.R
\name{smc2_control}
\alias{smc2_control}
\title{Control for SMC2}
\usage{
smc2_control(
  n_parameter_sets,
  degeneracy_threshold = 0.5,
  covariance_scaling = 0.5,
  progress = TRUE,
  save_trajectories = FALSE
)
}
\arguments{
\item{n_parameter_sets}{The number of replicate parameter sets to
simulate at once.}

\item{degeneracy_threshold}{The degeneracy threshold. Once the
effective sample size drops below \code{degeneracy_threshold * n_parameter_sets} the algorithm will rerun simulations from the
beginning of the data and use these to replenish the particles.}

\item{covariance_scaling}{A scaling factor to update variance
covariance matrix of sampled parameters by}

\item{progress}{Logical, indicating if a progress bar should be
displayed, using \code{\link[progress:progress_bar]{progress::progress_bar}}.}

\item{save_trajectories}{Logical, indicating if particle
trajectories should be saved during the simulation.}
}
\value{
A \code{smc2_control} object, which should not be modified once created.
}
\description{
Control for \link{smc2}. This function constructs a list of options and
does some basic validation to ensure that the options will work
well together. Do not manually change the values in this
object. Do not refer to any argument except \code{n_parameter_sets} by
position as the order of the arguments may change in future.
}
\examples{
mcstate::smc2_control(100)
}
