% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc2_parameters.R
\name{smc2_parameter}
\alias{smc2_parameter}
\title{Describe single pmcmc parameter}
\usage{
smc2_parameter(
  name,
  sample,
  prior,
  min = -Inf,
  max = Inf,
  discrete,
  integer = FALSE
)
}
\arguments{
\item{name}{Name for the parameter (a string)}

\item{sample}{A sampling function; it must take a single argument
representing the number of sampled to be returned. Typically
this will be a \code{r} probability function corresponding to the
sampling version of your prior (e.g., you might use \code{runif} and
\code{dunif} for \code{sample} and \code{prior}). If you provide \code{min}, \code{max}
or \code{integer} you \emph{must} ensure that your function returns
values that satisfy these constraints, as this is not (yet)
checked.}

\item{prior}{A prior function. It must be a function that takes a
single argument, being the value of this parameter.}

\item{min}{Optional minimum value for the parameter (otherwise
\code{-Inf}). If given, then \code{initial} must be at least this
value.}

\item{max}{Optional max value for the parameter (otherwise
\code{Inf}). If given, then \code{initial} must be at most this
value.}

\item{discrete}{Deprecated; use \code{integer} instead.}

\item{integer}{Logical, indicating if this parameter is an
integer. If \code{TRUE} then the parameter will be rounded
after a new parameter is proposed.}
}
\description{
Describe a single parameter for use within the SMC^2. Note that
the name is not set here, but will end up being naturally defined
when used with \code{\link{smc2_parameters}}, which collects
these together for use with \code{\link[=smc2]{smc2()}}.
}
\examples{
mcstate::smc2_parameter("a",
                        function(n) rnorm(n),
                        function(x) dnorm(n, log = TRUE))
}
