% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_proposal.R
\name{adaptive_proposal_control}
\alias{adaptive_proposal_control}
\title{Adaptive proposal control}
\usage{
adaptive_proposal_control(
  initial_scaling = 0.2,
  scaling_increment = 0.01,
  acceptance_target = 0.234,
  initial_vcv_weight = 1000,
  forget_rate = 0.2,
  forget_end = Inf,
  pre_diminish = 0
)
}
\arguments{
\item{initial_scaling}{The initial scaling of the variance
covariance matrix to be used to generate the multivariate normal
proposal for the random-walk Metropolis-Hasting algorithm.}

\item{scaling_increment}{The scaling increment which is added or
subtracted to the scaling factor of the variance-covariance
after each adaptive step.}

\item{acceptance_target}{The target for the fraction of proposals
that should be accepted (optimally) for the adaptive part of the
mixture model.}

\item{initial_weight}{Initial weight of the variance-covariance
matrix used to build the proposal of the random-walk. Higher
values translate into higher confidence of the initial
variance-covariance matrix and means that update from additional
samples will be slower.}

\item{adaptive_contribution}{The fractional contribution of the
adaptive part of the proposal. The proposal is based on a
mixture model, with the non adaptive part used for the proposal
with a probability 1-adaptive_contribution.}

\item{diminishing_adaptation}{Logical parameter, set to \code{TRUE} as default.
Indicates whether or not the adaptation of the scaling parameter
diminishes with each update. Users should switch this option to \code{FALSE}
with caution, as this will not ensure ergodicity of the MCMC algorithm.
Thus we would not recommend production of posterior samples with this
option set to \code{FALSE}, but users may want to use it for exploratory
purposes.}
}
\description{
Control for adaptive proposals, used in \link{pmcmc_control}
for deterministic models.
}
\details{
Efficient exploration of the parameter space during an MCMC might
be difficult when the target distribution is of high
dimensionality, especially if the target probability distribution
present a high degree of correlation.  Adaptive schemes are used
to "learn" on the fly the correlation structure by updating the
proposal distribution by recalculating the empirical
variance-covariance matrix and rescale it at each adaptive step of
the MCMC.

Our implementation of an adaptive MCMC algorithm is based on an
adaptation of the "BlkAdpMul" algorithm in Sherlock et
al. (ALGORITHM 6). The algorithm is based on a random-walk
Metropolis-Hasting algorithm where the proposal is a multi-variate
Normal distribution centered on the current point.

Sherlock C, Fearnhead P, Roberts GO (2010) The Random Walk
Metropolis: Linking Theory and Practice Through a Case
Study. Statistical Science 25:172–190.
}
