% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast.mcstate_pmcmc_list}
\title{Sample Grid Scan}
\usage{
forecast(x, ..., filter, n_sample_pars = 10, forecast_steps = 0, burn_in = 0)

\method{forecast}{mcstate_pmcmc_list}(x, ..., filter, n_sample_pars = 10, forecast_steps = 0, burn_in = 1)
}
\arguments{
\item{x}{Output of \code{\link{grid_search}}.}

\item{...}{Other arguments}

\item{filter}{A \code{particle_filter} object to run (the same
(as the one used to produce \code{scan_results})}

\item{n_sample_pars}{Number of parameters to be sampled. For a
\code{mcstate_scan} these will be pairs of parameters sampled in
proportion to their posterior probability. For a \code{mcstate_pmcmc_list}
this will be full parameter sets sampled evenly from the chains.
Integer. This will determine how many trajectories are returned.}

\item{forecast_steps}{Number of time steps being forecast. Default = 0}

\item{burn_in}{Amount of burn in to remove (if an \code{mcstate_pmcmc_list})}
}
\value{
\code{\link{list}}. First element (trajectories) is a 3
  dimensional array of trajectories (time, state, tranjectories). Second
  element (param_grid) is the parameters chosen when sampling from the
  \code{scan_results} grid and the third dimension (inputs) is a list of
  model inputs.
}
\description{
Take a grid search produced by \code{\link{grid_search}} and
sample \code{n_sample_pars} from the parameter grid uses based
on their probability. For each parameter pair chosen, run particle
filter with \code{num_particles}.
}
