% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic.R
\name{particle_deterministic}
\alias{particle_deterministic}
\title{Deterministic particle likelihood}
\description{
Create a deterministic version of the
\code{\link{particle_filter}} object, which runs a single
particle deterministically.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{model}}{The dust model generator being simulated (cannot be
re-bound)}

\item{\code{has_multiple_parameters}}{Logical, indicating if the
deterministic particle requires multiple parameter sets in a list
as inputs, and if it it will produce a vector of likelihoods
the same length (read only).  The parameter sets may or may
not use the same data (see \code{has_multiple_data}).}

\item{\code{has_multiple_data}}{Logical, indicating if the deterministic
particle simultaneously calculates the likelihood for multiple
parameter sets (read only). If \code{TRUE}, \code{has_multiple_parameters}
will always be \code{TRUE}.}

\item{\code{n_parameters}}{The number of parameter sets used by this
deterministic particle (read only).  The returned vector of
likelihoods will be this length, and if \code{has_multiple_parameters}
is \code{FALSE} this will be 1.}

\item{\code{n_data}}{The number of data sets used by this deterministic
particle (read only).  This will either be 1 or the same value as
\code{n_parameters}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-particle_deterministic-new}{\code{particle_deterministic$new()}}
\item \href{#method-particle_deterministic-run}{\code{particle_deterministic$run()}}
\item \href{#method-particle_deterministic-run_begin}{\code{particle_deterministic$run_begin()}}
\item \href{#method-particle_deterministic-state}{\code{particle_deterministic$state()}}
\item \href{#method-particle_deterministic-history}{\code{particle_deterministic$history()}}
\item \href{#method-particle_deterministic-restart_state}{\code{particle_deterministic$restart_state()}}
\item \href{#method-particle_deterministic-inputs}{\code{particle_deterministic$inputs()}}
\item \href{#method-particle_deterministic-set_n_threads}{\code{particle_deterministic$set_n_threads()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-new"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-new}{}}}
\subsection{Method \code{new()}}{
Create the particle filter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$new(
  data,
  model,
  compare,
  index = NULL,
  initial = NULL,
  constant_log_likelihood = NULL,
  n_threads = 1L,
  n_parameters = NULL,
  stochastic_schedule = NULL,
  ode_control = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The data set to be used for the particle filter,
created by \code{\link[=particle_filter_data]{particle_filter_data()}}. This is essentially
a \code{\link[=data.frame]{data.frame()}} with at least columns \code{time_start}
and \code{time_end}, along with any additional data used in the
\code{compare} function, and additional information about how your
steps relate to time.}

\item{\code{model}}{A stochastic model to use.  Must be a
\code{dust_generator} object.}

\item{\code{compare}}{A comparison function.  Must take arguments
\code{state}, \code{observed} and \code{pars} as arguments (though the arguments
may have different names). \code{state} is the simulated model state
(a matrix with as many rows as there are state variables and as
many columns as there are particles, \code{data}
is a \code{list} of observed data corresponding to the current
time's row in the \code{data} object provided here in the
constructor.  \code{pars} is any additional parameters passed
through to the comparison function (via the \code{pars}
argument to \verb{$run}).}

\item{\code{index}}{An index function. This is used to compute the
"interesting" indexes of your model. It must be a function of
one argument, which will be the result of calling the
\verb{$info()} method on your model. It should return a list
with elements \code{run} (indices to return at the end of each
run, passed through to your compare function) and \code{state}
(indices to return if saving state). These indices can overlap
but do not have to. This argument is optional but using it will
likely speed up your simulation if you have more than a few
states as it will reduce the amount of memory copied back and
forth.}

\item{\code{initial}}{A function to generate initial conditions. If
given, then this function must accept 3 arguments: \code{info}
(the result of calling \verb{$info()} as for \code{index}),
\code{n_particles} (the number of particles that the particle
filter is using) and \code{pars} (parameters passed in in the
\verb{$run} method via the \code{pars} argument).  It
must return a list, which can have the elements \code{state}
(initial model state, passed to the particle filter - either a
vector or a matrix, and overriding the initial conditions
provided by your model) and \code{time} (the initial time,
overriding the first time step of your data - this must occur within
your first epoch in your \code{data} provided to the
constructor, i.e., not less than the first element of
\code{time_start} and not more than \code{time_end}). Your function
can also return a vector or matrix of \code{state} and not alter
the starting time, which is equivalent to returning
\code{list(state = state, time = NULL)}.
(TODO: this no longer is allowed, and the docs might be out of date?)}

\item{\code{constant_log_likelihood}}{An optional function, taking the
model parameters, that computes the constant part of the
log-likelihood value (if any).  You can use this where your
likelihood depends both on the time series (via \code{data}) but also
on some non-temporal data.  You should bind any non-parameter
dependencies into this closure.  This is applied at the
beginning of the filter run, so represents the initial
condition of the marginal log likelihood value propagated by
the process.}

\item{\code{n_threads}}{Number of threads to use when running the
simulation. Defaults to 1, and should not be set higher than the
number of cores available to the machine. This currently has no
effect as the simulation will be run in serial on a single
particle for now.}

\item{\code{n_parameters}}{Number of parameter sets required.  This, along
with \code{data}, controls the interpretation of how the deterministic
particle, and importantly will add an additional dimension to
most outputs (scalars become vectors, vectors become matrices etc).}

\item{\code{stochastic_schedule}}{Vector of times to perform stochastic
updates, for continuous time models. Note that despite the name,
these will be applied deterministically (i.e., replacing the
stochastic draw with its expectation).}

\item{\code{ode_control}}{Tuning control for the ODE stepper, for
continuous time (ODE) models}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-run"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-run}{}}}
\subsection{Method \code{run()}}{
Run the deterministic particle filter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$run(
  pars = list(),
  save_history = FALSE,
  save_restart = NULL,
  min_log_likelihood = -Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{A list representing parameters. This will be passed as
the \code{pars} argument to your model, to your \code{compare}
function, and (if using) to your \code{initial} function. It must
be an R list (not vector or \code{NULL}) because that is what a
dust model currently requires on initialisation or \verb{$reset} - we
may relax this later. You may want to put your observation and
initial parameters under their own keys (e.g.,
\code{pars$initial$whatever}), but this is up to you. Extra keys
are silently ignored by dust models.}

\item{\code{save_history}}{Logical, indicating if the history of all
particles should be saved. If saving history, then it can be
queried later with the \verb{$history} method on the object.}

\item{\code{save_restart}}{An integer vector of time points to save
restart infomation for. Not currently supported.}

\item{\code{min_log_likelihood}}{Not currently supported, exists to match
the inteface with \link{particle_filter}. Providing a value
larger than -Inf will cause an error.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A single numeric value representing the log-likelihood
(\code{-Inf} if the model is impossible)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-run_begin"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-run_begin}{}}}
\subsection{Method \code{run_begin()}}{
Begin a deterministic run. This is part of the
"advanced" interface; typically you will want to use \verb{$run()}
which provides a user-facing wrapper around
this function. Once created with \verb{$run_begin()}, you should take
as many steps as needed with \verb{$step()}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$run_begin(
  pars,
  save_history = FALSE,
  save_restart = NULL,
  min_log_likelihood = -Inf
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pars}}{A list representing parameters. See \verb{$run_many()}
for details (and \emph{not} \verb{$run()})}

\item{\code{save_history}}{Logical, indicating if the history of all
particles should be saved. See \verb{$run()} for details.}

\item{\code{save_restart}}{Times to save restart state at. See \verb{$run()} for
details.}

\item{\code{min_log_likelihood}}{Not currently supported, exists to match
the inteface with \link{particle_filter}. Providing a value
larger than -Inf will cause an error.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{particle_deterministic_state}, with methods
\code{step} and \code{end}. This interface is still subject to change.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-state"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-state}{}}}
\subsection{Method \code{state()}}{
Extract the current model state, optionally filtering.
If the model has not yet been run, then this method will throw an
error. Returns a matrix with the number of rows being the number of
model states, and the number of columns being the number of
particles.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$state(index_state = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index_state}}{Optional vector of states to extract}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-history"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-history}{}}}
\subsection{Method \code{history()}}{
Extract the particle trajectories. Requires that
the model was run with \code{save_history = TRUE}, which does
incur a performance cost. This method will throw an error if
the model has not run, or was run without \code{save_history = TRUE}. Returns a 3d array with dimensions corresponding to (1)
model state, filtered by \code{index$run} if provided, (2)
particle (following \code{index_particle} if provided), (3)
time point.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$history(index_particle = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index_particle}}{Optional vector of particle indices to return.
If \code{NULL} we return all particles' histories.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-restart_state"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-restart_state}{}}}
\subsection{Method \code{restart_state()}}{
Return the full particle filter state at points back in time
that were saved with the \code{save_restart} argument to
\verb{$run()}. If available, this will return a 3d array, with
dimensions representing (1) particle state, (2) particle index,
(3) time point. If multiple parameters are used then returns a 4d array,
with dimensions representing (1) particle state, (2) particle index,
(3) parameter, (4) time point. This could be quite large, especially
if you are using the \code{index} argument to create the particle filter
and return a subset of all state generally. In the stochastic version,
this is different the saved trajectories returned by \verb{$history()}
because earlier saved state is not filtered by later filtering,
but in the deterministic model we run with a single particle so it
\emph{is} the same.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$restart_state(
  index_particle = NULL,
  save_restart = NULL,
  restart_match = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index_particle}}{Optional vector of particle indices to return.
If \code{NULL} we return all particles' states. Practically because the
only valid value of index_particle is "1", this has no effect and
it is included primarily for compatibility with the stochastic
interface.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-inputs"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-inputs}{}}}
\subsection{Method \code{inputs()}}{
Return a list of inputs used to configure the deterministic particle
filter. These correspond directly to the argument names for the
constructor and are the same as the input arguments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$inputs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-particle_deterministic-set_n_threads"></a>}}
\if{latex}{\out{\hypertarget{method-particle_deterministic-set_n_threads}{}}}
\subsection{Method \code{set_n_threads()}}{
Set the number of threads used by the particle filter (and dust
model) after creation. This can be used to allocate additional
(or subtract excess) computing power from the deterministic filter
Returns (invisibly) the previous value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{particle_deterministic$set_n_threads(n_threads)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_threads}}{The new number of threads to use. You may want to
wrap this argument in \code{\link[dust:dust_openmp_threads]{dust::dust_openmp_threads()}} in order to
verify that you can actually use the number of threads
requested (based on environment variables and OpenMP support).}
}
\if{html}{\out{</div>}}
}
}
}
