% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc.R
\name{pmcmc}
\alias{pmcmc}
\title{Run a pmcmc sampler}
\usage{
pmcmc(
  mcmc_range,
  lprior_funcs,
  filter,
  n_mcmc,
  proposal_kernel,
  output_proposals = FALSE,
  n_chains = 1
)
}
\arguments{
\item{mcmc_range}{\code{data.frame} detailing parameters to sample.
Must contain columns 'name' (parameter names), 'init' (initial values),
'min' (minimum values), 'max' (maximum values),
'discrete (boolean indicating whether a discrete quantity)' and
'target'}

\item{lprior_funcs}{functions to calculate log prior for each parameter.
A named list for each parameter listed in \code{pars_to_sample}.
Each value must be a function which takes named parameter vector as
  input, returns a single numeric which is log of the prior probability.}

\item{filter}{A particle filter to sample}

\item{n_mcmc}{Number of MCMC steps}

\item{proposal_kernel}{named matrix of proposal covariance for parameters}

\item{output_proposals}{Logical indicating whether proposed parameter
jumps should be output along with results}

\item{n_chains}{Number of chains to run}
}
\value{
an mcmc object containing
- List of inputs
- Matrix of accepted parameter samples, rows = iterations
  as well as log prior, (particle filter estimate of) log likelihood
  and log posterior
}
\description{
This is a basic Metropolis-Hastings MCMC sampler.
The \code{filter} is run with a set of parameters to evaluate the
likeihood. A new set of parameters is proposed, and these likelihoods are
compared, jumping with probability equal to their ratio. This is repeated
for \code{n_mcmc} proposals, \code{n_chains} independent times.
}
\details{
Run a pmcmc sampler
}
