% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_parameters_nested.R
\name{pmcmc_parameters_nested}
\alias{pmcmc_parameters_nested}
\title{pmcmc_parameters}
\description{
Construct parameters for use with
\code{\link[=pmcmc]{pmcmc()}}. This creates a utility object that is used
internally to work with parameters. Most users only need to
construct this object, but see the examples for how it can be
used.
}
\examples{
#Construct an object with two parameters:
pars <- mcstate::pmcmc_parameters$new(
  list(mcstate::pmcmc_parameter("a", 0.1, min = 0, max = 1,
                                prior = function(a) log(a)),
       mcstate::pmcmc_parameter("b", 0, prior = dnorm)),
  matrix(c(1, 0.5, 0.5, 2), 2, 2))

# Initial parameters
p <- pars$initial()
p

# Propose a new parameter point
pars$propose(p)

# Information about parameters:
pars$names()
pars$summary()

# Compute prior
pars$prior(p)

# Transform data for your model
pars$model(p)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pmcmc_parameters_nested$new()}}
\item \href{#method-names}{\code{pmcmc_parameters_nested$names()}}
\item \href{#method-populations}{\code{pmcmc_parameters_nested$populations()}}
\item \href{#method-summary}{\code{pmcmc_parameters_nested$summary()}}
\item \href{#method-initial}{\code{pmcmc_parameters_nested$initial()}}
\item \href{#method-prior}{\code{pmcmc_parameters_nested$prior()}}
\item \href{#method-propose}{\code{pmcmc_parameters_nested$propose()}}
\item \href{#method-model}{\code{pmcmc_parameters_nested$model()}}
\item \href{#method-fix}{\code{pmcmc_parameters_nested$fix()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create the pmcmc_parameters object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$new(
  parameters,
  proposal_varied = NULL,
  proposal_fixed = NULL,
  populations = NULL,
  transform = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A \code{list} of
\link{pmcmc_parameter} objects, each of which describe a
single parameter in your model. If \code{parameters} is named, then
these names must match the \verb{$name} element of each parameter is
used (this is verified).}

\item{\code{populations}}{Specifies the names of the different populations
that the parameters vary according to. Assumes that varied parameters
are estimated based on representative samples drawn from the respective
populations. Ignored if no varied parameters are included.}

\item{\code{transform}}{An optional transformation function to apply
to your parameter vector immediately before passing it to the
model function. If not given, then \link{as.list} is
used, as dust models require this. However, if you need to
generate derived parameters from those being actively sampled
you can do arbitrary transformations here.}

\item{\code{proposal}}{A square proposal distribution corresponding to the
variance-covariance matrix of a multivariate gaussian distribution
used to generate new parameters. It must have the same number of
rows and columns as there are elements in \code{parameters}, and if
named the names must correspond exactly to the names in
\code{parameters}. Because it corresponds to a variance-covariance
matrix it must be symmetric and positive definite.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-names"></a>}}
\if{latex}{\out{\hypertarget{method-names}{}}}
\subsection{Method \code{names()}}{
Return the names of the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-populations"></a>}}
\if{latex}{\out{\hypertarget{method-populations}{}}}
\subsection{Method \code{populations()}}{
Return the names of the populations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$populations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Return a \code{data.frame} with information about
parameters (name, min, max, discrete, fixed or varied).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$summary(population = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{population}}{For parameter sets including varied parameters,
\code{population} specifies which population to summarise. If \code{NULL} then
returns summary of each population as a list.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-initial"></a>}}
\if{latex}{\out{\hypertarget{method-initial}{}}}
\subsection{Method \code{initial()}}{
Return the initial parameter values as a named numeric
vector or named matrix for multiple populations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$initial()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prior"></a>}}
\if{latex}{\out{\hypertarget{method-prior}{}}}
\subsection{Method \code{prior()}}{
Compute the prior(s) for a parameter vector/matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$prior(theta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{a parameter matrix with columns in the same order as
\verb{$names()} and rows in the same order as \verb{$populations()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-propose"></a>}}
\if{latex}{\out{\hypertarget{method-propose}{}}}
\subsection{Method \code{propose()}}{
Propose a new parameter vector given a current parameter
vector. This proposes a new parameter vector/matrix given your current
vector and the variance-covariance matrix of your proposal
kernel, discretises any discrete values, and reflects bounded
parameters until they lie within \code{min}:\code{max}. Returns matrix if any
varied parameters are included, otherwise vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$propose(
  theta,
  scale = 1,
  type = c("fixed", "varied", "both")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{a parameter vector in the same order as your
parameters were defined in (see \verb{$names()} for that order.}

\item{\code{scale}}{an optional scaling factor to apply to the
proposal distribution. This may be useful in sampling starting
points. The parameter is equivalent to a multiplicative factor
applied to the variance covariance matrix.}

\item{\code{type}}{specifies which type of parameters should be proposed,
either fixed parameters only ("fixed"), varied only ("varied"), or
both ("both") types. For 'fixed' and 'varied' then parameters of the
other type is returned equal to corresponding value in \code{theta}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model"></a>}}
\if{latex}{\out{\hypertarget{method-model}{}}}
\subsection{Method \code{model()}}{
Apply the model transformation function to a parameter
vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$model(theta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{a parameter vector in the same order as your
parameters were defined in (see \verb{$names()} for that order.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fix"></a>}}
\if{latex}{\out{\hypertarget{method-fix}{}}}
\subsection{Method \code{fix()}}{
Set some parameters to fixed values. Use this to
reduce the dimensionality of your system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pmcmc_parameters_nested$fix(fixed)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fixed}}{a named vector of parameters to fix}
}
\if{html}{\out{</div>}}
}
}
}
