% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_tools.R
\name{pmcmc_tidy_trajectories}
\alias{pmcmc_tidy_trajectories}
\title{Convert sampled trajectories from \code{\link[=pmcmc]{pmcmc()}} to a Tidy (long) format}
\usage{
pmcmc_tidy_trajectories(
  object,
  summarise = FALSE,
  quantiles = c(0.025, 0.5, 0.975),
  summary_function = NULL
)
}
\arguments{
\item{object}{Results of running \code{\link[=pmcmc]{pmcmc()}}.}

\item{summarise}{Logical indicating whether the trajectories should be
marginalised over particles to produce summary statistics. Default is
\code{FALSE}.}

\item{quantiles}{Optional vector giving quantiles to output if \code{summarise}
is TRUE defaults to c(0.025, 0.5, 0.975) corresponding to 2.5-, 50-, and
97.5-percentiles}

\item{summary_function}{Optional function to calculated user-defined
summary statistics - defaults to calculating mean, median, 2.5\% and 97.5\%
percentiles}
}
\description{
Convert sampled trajectories from \code{\link[=pmcmc]{pmcmc()}} to a Tidy (long) format,
and optionally marginalise over particles to produce summary statistics.
Helper function useful for ggplotters.
}
