% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcmc_tools.R
\name{pmcmc_tidy_trajectories}
\alias{pmcmc_tidy_trajectories}
\title{Convert sampled trajectories from \code{\link[=pmcmc]{pmcmc()}} to a Tidy (long) format}
\usage{
pmcmc_tidy_trajectories(object, summarise = FALSE, summary = NULL)
}
\arguments{
\item{object}{Results of running \code{\link[=pmcmc]{pmcmc()}}.}

\item{summarise}{Logical indicating whether the trajectories should be
marginalised over particles to produce summary statistics. Default is
\code{FALSE}.}

\item{summary}{Optional function to calculate user-defined
summary statistics. A user-supplied function should take a single argument
(which will be a numeric vector of length \code{n_particles} corresponding to a
single state, for a single population, at a single time point) and return a
named numeric vector of summary statistics. The default calculates the mean,
2.5-, 50-, and 97.5-percentiles}
}
\description{
Convert sampled trajectories from \code{\link[=pmcmc]{pmcmc()}} to a Tidy (long) format,
and optionally marginalise over particles to produce summary statistics.
Helper function useful for ggplotters.
}
