% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc2_parameters.R
\name{smc2_parameters}
\alias{smc2_parameters}
\title{pmcmc_parameters}
\description{
Construct parameters for use with
\code{\link[=smc]{smc()}}. This creates a utility object that is used
internally to work with parameters. Most users only need to
construct this object, but see the examples for how it can be
used.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{smc2_parameters$new()}}
\item \href{#method-sample}{\code{smc2_parameters$sample()}}
\item \href{#method-names}{\code{smc2_parameters$names()}}
\item \href{#method-summary}{\code{smc2_parameters$summary()}}
\item \href{#method-prior}{\code{smc2_parameters$prior()}}
\item \href{#method-propose}{\code{smc2_parameters$propose()}}
\item \href{#method-model}{\code{smc2_parameters$model()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create the smc2_parameters object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smc2_parameters$new(parameters, transform = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{A \code{list} of
\link{smc2_parameter} objects, each of which describe a
single parameter in your model. If \code{parameters} is named, then
these names must match the \verb{$name} element of each parameter is
used (this is verified).}

\item{\code{transform}}{An optional transformation function to apply
to your parameter vector immediately before passing it to the
model function. If not given, then \link{as.list} is
used, as dust models require this. However, if t you need to
generate derived parameters from those being actively sampled
you can do arbitrary transformations here.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
Create \code{n} independent random parameter vectors (as a
matrix with \code{n} rows)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smc2_parameters$sample(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of replicate parameter sets to draw}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-names"></a>}}
\if{latex}{\out{\hypertarget{method-names}{}}}
\subsection{Method \code{names()}}{
Return the names of the parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smc2_parameters$names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-summary"></a>}}
\if{latex}{\out{\hypertarget{method-summary}{}}}
\subsection{Method \code{summary()}}{
Return a \code{data.frame} with information about
parameters (name, min, max, and discrete).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smc2_parameters$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-prior"></a>}}
\if{latex}{\out{\hypertarget{method-prior}{}}}
\subsection{Method \code{prior()}}{
Compute the prior for a parameter vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smc2_parameters$prior(theta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{a parameter vector in the same order as your
parameters were defined in (see \verb{$names()} for that order.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-propose"></a>}}
\if{latex}{\out{\hypertarget{method-propose}{}}}
\subsection{Method \code{propose()}}{
Propose a new parameter vector given a current parameter
vector and variance covariance matrix. After proposal, this discretises
any discrete values, and reflects bounded parameters until they lie
within \code{min}:\code{max}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smc2_parameters$propose(theta, vcv)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{a parameter vector in the same order as your
parameters were defined in (see \verb{$names()} for that order).}

\item{\code{vcv}}{the variance covariance matrix for the proposal; must
be square and have a number of rows and columns equal to the
number of parameters, in the same order as \code{theta}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-model"></a>}}
\if{latex}{\out{\hypertarget{method-model}{}}}
\subsection{Method \code{model()}}{
Apply the model transformation function to a parameter
vector.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{smc2_parameters$model(theta)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{theta}}{a parameter vector in the same order as your
parameters were defined in (see \verb{$names()} for that order.}
}
\if{html}{\out{</div>}}
}
}
}
